refactor(): returns buildUiSchema
