+++ packages/common/src/discussions/_internal/DiscussionUiSchemaSettings.ts
@@ -1,4 +1,6 @@
-import { IUiSchema } from "../../core";
+import { IHubDiscussion } from "../../core/types";
+import { IArcGISContext } from "../../ArcGISContext";
+import { IUiSchema } from "../../core/schemas/types";
 
 /**
  * @private
@@ -6,31 +8,37 @@ import { IUiSchema } from "../../core";
  * defines how the schema properties should be
  * rendered in the Discussions settings experience
  */
-export const uiSchema: IUiSchema = {
-  type: "Layout",
-  elements: [
-    {
-      type: "Section",
-      labelKey: "{{i18nScope}}.sections.settings.label",
-      elements: [
-        {
-          labelKey: "{{i18nScope}}.fields.discussable.label",
-          scope: "/properties/isDiscussable",
-          type: "Control",
-          options: {
-            control: "hub-field-input-radio",
-            labels: [
-              "{{{{i18nScope}}.fields.discussable.enabled.label:translate}}",
-              "{{{{i18nScope}}.fields.discussable.disabled.label:translate}}",
-            ],
-            descriptions: [
-              "{{{{i18nScope}}.fields.discussable.enabled.description:translate}}",
-              "{{{{i18nScope}}.fields.discussable.disabled.description:translate}}",
-            ],
-            icons: ["speech-bubbles", "circle-disallowed"],
+export const buildUiSchema = async (
+  i18nScope: string,
+  entity: IHubDiscussion,
+  context: IArcGISContext
+): Promise<IUiSchema> => {
+  return {
+    type: "Layout",
+    elements: [
+      {
+        type: "Section",
+        labelKey: `${i18nScope}.sections.settings.label`,
+        elements: [
+          {
+            labelKey: `${i18nScope}.fields.discussable.label`,
+            scope: "/properties/isDiscussable",
+            type: "Control",
+            options: {
+              control: "hub-field-input-radio",
+              labels: [
+                `{{${i18nScope}.fields.discussable.enabled.label:translate}}`,
+                `{{${i18nScope}.fields.discussable.disabled.label:translate}}`,
+              ],
+              descriptions: [
+                `{{${i18nScope}.fields.discussable.enabled.description:translate}}`,
+                `{{${i18nScope}.fields.discussable.disabled.description:translate}}`,
+              ],
+              icons: ["speech-bubbles", "circle-disallowed"],
+            },
           },
-        },
-      ],
-    },
-  ],
+        ],
+      },
+    ],
+  };
 };
